computerB-VR/requirements.txt
numpy>=1.24
opencv-python-headless>=4.8
websockets>=12.0
PyYAML>=6.0
websocket-client>=1.6
